rename commands folders in cmd/kafeman
