+++ cmd/kafeman/completion_cmd/completion.go
