+++ cmd/kafeman/global_config/global_config.go
@@ -1,4 +1,4 @@
-package global_config
+package common
 
 import (
 	"os"
